created a list container to remove the logic from the home component
